Separate out each problem into its own self class
