+++ ruby/Dynamic Programming/dynamic_programming_test.rb
@@ -1,0 +1,7 @@
+def test(method_name, method_test)
+  if method_test
+    p "#{method_name} OK"
+  else
+    p "#{method_name} FAIL"
+  end
+end
